Turning this into a rust debug library
