+++ embedded-rust-debugger/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name	= "embedded-rust-debugger"
+name	= "rust-debug"
 version = "0.1.0"
 authors = ["Niklas <nicke.l@telia.com>"]
 license = "MIT OR Apache-2.0"
