+++ embedded-rust-debugger/src/call_stack.rs
@@ -8,7 +8,7 @@
 
 use gimli::DebugFrame;
 
-use crate::rust_debug::memory_and_registers::MemoryAndRegisters;
+use crate::memory_and_registers::MemoryAndRegisters;
 
 use gimli::{
     Reader,
