--- embedded-rust-debugger/src/commands/mod.rs
@@ -1,12 +1,0 @@
-pub mod debug_request;
-pub mod debug_response;
-pub mod debug_event;
-pub mod commands;
-
-
-pub enum Command {
-    Request(debug_request::DebugRequest),
-    Response(debug_response::DebugResponse),
-    Event(debug_event::DebugEvent),
-}
-
